requirements.txt
# Core dependencies
Flask>=2.0.1
Flask-SQLAlchemy>=2.5.1
Flask-Migrate>=3.1.0
Flask-RESTful>=0.3.9
Flask-SocketIO>=5.1.1

# System monitoring
psutil>=5.8.0
watchdog>=2.1.9

# GPU monitoring
nvidia-ml-py3>=7.352.0; sys_platform == 'linux'
wmi>=1.5.1; sys_platform == 'win32'
pythonnet>=3.0.1; sys_platform == 'win32'  # For OpenHardwareMonitor
pyamdgpuinfo>=2.1.5; sys_platform == 'linux'

# Database
SQLAlchemy>=1.4.23
alembic>=1.7.1

# Utils
PyYAML==6.0.1
requests==2.31.0
python-dotenv>=0.19.0
pydantic>=1.8.2

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1

# Monitoring
prometheus-client>=0.11.0

# Type hints
typing-extensions>=3.10.0

# Development
black>=21.5b2
flake8>=3.9.2
mypy>=0.910
